increase flake8 line limit && fix noqa on some files
Flake was also complaining about the length of comments and strings which are displayed properly in the docs. And blake is enough to ensure that code fits in the docs.

Line limit decreased because some strings were still not fitting.

And also `flake: noqa: E402` doesn't work -- that line disables all quality assurance for that file so instead `per-file-ignores` are added.